fix: move all editDialogModal into train
