+++ cypress/integration/CreateModels/WaitVsNoWaitActions.spec.js
@@ -6,7 +6,6 @@
 import * as models from '../../support/Models'
 import * as modelPage from '../../support/components/ModelPage'
 import * as actions from '../../support/Actions'
-import * as editDialogModal from '../../support/components/EditDialogModal'
 import * as train from '../../support/Train'
 
 describe('CreateModels', () => {
@@ -25,12 +24,12 @@ describe('CreateModels', () => {
     train.CreateNewTrainDialog()
 
     train.TypeYourMessage('Duck')
-    editDialogModal.ClickScoreActionsButton()
+    train.ClickScoreActionsButton()
     train.SelectAction('Ducks say quack!')
     train.SelectAction('Which animal would you like?')
 
     train.TypeYourMessage('Fish')
-    editDialogModal.ClickScoreActionsButton()
+    train.ClickScoreActionsButton()
     train.SelectAction('Fish just swim.')
     train.SelectAction('Which animal would you like?')
