+++ cypress/integration/EditAndBranching/TagAndFrog.spec.js
@@ -7,7 +7,6 @@ import * as models from '../../support/Models'
 import * as modelPage from '../../support/components/ModelPage'
 import * as scorerModal from '../../support/components/ScorerModal'
 import * as train from '../../support/Train'
-import * as editDialogModal from '../../support/components/EditDialogModal'
 import * as helpers from '../../support/Helpers'
 
 describe('Tag and Frog - Edit And Branching', () => {
@@ -26,26 +25,26 @@ describe('Tag and Frog - Edit And Branching', () => {
   context('Edit Train Dialog', () => {
     it('Should remove two entity labels from alternative input', () => {
       train.EditTraining('This is Tag.', 'This is Tag.', 'Hi')
-      editDialogModal.SelectChatTurnExactMatch('This is Tag.')
+      train.SelectChatTurnExactMatch('This is Tag.')
 
-      editDialogModal.VerifyEntityLabelWithinSpecificInput([textEntityPairs[0]], 0)
-      editDialogModal.VerifyEntityLabelWithinSpecificInput(textEntityPairs, 1)
-      editDialogModal.VerifyEntityLabelWithinSpecificInput(textEntityPairs, 2)
+      train.VerifyEntityLabelWithinSpecificInput([textEntityPairs[0]], 0)
+      train.VerifyEntityLabelWithinSpecificInput(textEntityPairs, 1)
+      train.VerifyEntityLabelWithinSpecificInput(textEntityPairs, 2)
 
-      editDialogModal.RemoveEntityLabel('Tag', 'multi', 1)
-      editDialogModal.RemoveEntityLabel('Frog', 'multi', 2)
+      train.RemoveEntityLabel('Tag', 'multi', 1)
+      train.RemoveEntityLabel('Frog', 'multi', 2)
     })
 
     
     it('Should verify user cannot submit changes without accepting auto-re-labling of those two entity labels in the alternative input', () => {
-      editDialogModal.ClickSubmitChangesButton()
-      editDialogModal.VerifyEntityLabeledDifferentPopupAndClose(textEntityPairs)
-      editDialogModal.ClickSubmitChangesButton()
-      editDialogModal.VerifyEntityLabeledDifferentPopupAndAccept(textEntityPairs)
+      train.ClickSubmitChangesButton()
+      train.VerifyEntityLabeledDifferentPopupAndClose(textEntityPairs)
+      train.ClickSubmitChangesButton()
+      train.VerifyEntityLabeledDifferentPopupAndAccept(textEntityPairs)
 
-      editDialogModal.VerifyEntityLabeledDifferentPopupAndClose(textEntityPairs)
-      editDialogModal.ClickSubmitChangesButton()
-      editDialogModal.VerifyEntityLabeledDifferentPopupAndAccept(textEntityPairs)
+      train.VerifyEntityLabeledDifferentPopupAndClose(textEntityPairs)
+      train.ClickSubmitChangesButton()
+      train.VerifyEntityLabeledDifferentPopupAndAccept(textEntityPairs)
 
       train.AbandonDialog()
     })
