+++ cypress/integration/EditAndBranching/VerifyEditTrainingControlsAndLabels.spec.js
@@ -7,7 +7,6 @@ import * as models from '../../support/Models'
 import * as modelPage from '../../support/components/ModelPage'
 import * as scorerModal from '../../support/components/ScorerModal'
 import * as train from '../../support/Train'
-import * as editDialogModal from '../../support/components/EditDialogModal'
 
 describe('EditAndBranching', () => {
   it('Verify Edit Training Controls And Labels', () => {
@@ -17,18 +16,18 @@ describe('EditAndBranching', () => {
     train.EditTraining('My name is David.', 'My name is Susan.', 'Hello $name')
     train.CaptureOriginalChatMessages()
 
-    editDialogModal.VerifyCloseButtonLabel()
-    editDialogModal.VerifyDeleteButtonLabel()
+    train.VerifyCloseButtonLabel()
+    train.VerifyDeleteButtonLabel()
 
-    editDialogModal.VerifyThereAreNoChatEditControls('My name is David.', 'Hello Susan')
+    train.VerifyThereAreNoChatEditControls('My name is David.', 'Hello Susan')
     train.SelectAndVerifyEachChatTurn()
 
     train.BranchChatTurn('My name is Susan.', 'I am Groot')
-    editDialogModal.VerifySaveBranchButtonLabel()
-    editDialogModal.VerifyAbandonBranchButtonLabel()
+    train.VerifySaveBranchButtonLabel()
+    train.VerifyAbandonBranchButtonLabel()
 
-    editDialogModal.VerifyThereAreNoChatEditControls('I am Groot', 'Hello David')
-    editDialogModal.AbandonBranchChanges()
+    train.VerifyThereAreNoChatEditControls('I am Groot', 'Hello David')
+    train.AbandonBranchChanges()
 
     train.EditTraining('My name is David.', 'My name is Susan.', 'Hello $name')
     train.VerifyOriginalChatMessages()
