+++ cypress/integration/EndSession/AddEndSessionAction.spec.js
@@ -6,7 +6,6 @@
 import * as models from '../../support/Models'
 import * as modelPage from '../../support/components/ModelPage'
 import * as train from '../../support/Train'
-import * as editDialogModal from '../../support/components/EditDialogModal'
 
 describe('EditAndBranching', () => {
   it('Add End Session Action', () => {
@@ -17,12 +16,12 @@ describe('EditAndBranching', () => {
     cy.WaitForTrainingStatusCompleted()
 
     train.EditTraining('fly to sydney', 'coach', "enjoy your trip. you are booked on Qantas")
-    editDialogModal.ClickScoreActionsButton()
-    editDialogModal.SelectChatTurnExactMatch('enjoy your trip. you are booked on Qantas', 1)
+    train.ClickScoreActionsButton()
+    train.SelectChatTurnExactMatch('enjoy your trip. you are booked on Qantas', 1)
     train.SelectEndSessionAction('0')
 
-    editDialogModal.VerifyScoreActionsButtonIsMissing()
-    editDialogModal.VerifyTypeYourMessageIsMissing()
-    editDialogModal.ClickSaveCloseButton()
+    train.VerifyScoreActionsButtonIsMissing()
+    train.VerifyTypeYourMessageIsMissing()
+    train.ClickSaveCloseButton()
   })
 })
