+++ cypress/integration/ErrorHandling/ActionUnavailable.spec.js
@@ -6,7 +6,6 @@
 import * as models from '../../support/Models'
 import * as modelPage from '../../support/components/ModelPage'
 import * as train from '../../support/Train'
-import * as editDialogModal from '../../support/components/EditDialogModal'
 import * as common from '../../support/Common'
 import * as helpers from '../../support/Helpers'
 
@@ -29,29 +28,29 @@ describe('Action Unavailable - ErrorHandling', () => {
 
     it('Should add user turn "Joe" and label it as the "name" entity', () => {
       train.TypeYourMessage('Joe')
-      editDialogModal.LabelTextAsEntity('Joe', 'name')
+      train.LabelTextAsEntity('Joe', 'name')
     })
 
     it('Should Score Actions to train the Bot to respond with "Hello Joe"', () => {
-      editDialogModal.ClickScoreActionsButton()
+      train.ClickScoreActionsButton()
       train.SelectAction('Hello Joe')
     })
 
     it('Should introduce an error in the Bot response by removing the entity label from "Joe"', () => {
-      editDialogModal.SelectChatTurnExactMatch('Joe')
-      editDialogModal.RemoveEntityLabel('Joe', 'name')
-      editDialogModal.ClickSubmitChangesButton()
-      editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+      train.SelectChatTurnExactMatch('Joe')
+      train.RemoveEntityLabel('Joe', 'name')
+      train.ClickSubmitChangesButton()
+      train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
     })
 
     it('Should verify the specifics of the error', () => {
-      editDialogModal.SelectChatTurnStartsWith('Hello')
-      editDialogModal.VerifyErrorMessage('Action is unavailable')
-      editDialogModal.VerifyChatTurnIsAnExactMatch('Hello [$name]', 2, 1)
+      train.SelectChatTurnStartsWith('Hello')
+      train.VerifyErrorMessage('Action is unavailable')
+      train.VerifyChatTurnIsAnExactMatch('Hello [$name]', 2, 1)
     })
 
     it('Should save the training with errors', () => {
-      editDialogModal.ClickSaveCloseButton()
+      train.ClickSaveCloseButton()
       modelPage.VerifyIncidentTriangleForTrainDialogs()
       train.VerifyErrorsFoundInTraining(`Joe`, 'Joe', "Hello $name")
     })
@@ -60,21 +59,21 @@ describe('Action Unavailable - ErrorHandling', () => {
   context('Edit Dialog - Validate Errors and Fix Them', () => {
     it('Should edit the training and verify it has errors', () => {
       train.EditTraining(`Joe`, 'Joe', "Hello $name")
-      editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
-      editDialogModal.SelectChatTurnStartsWith('Hello')
-      editDialogModal.VerifyErrorMessage('Action is unavailable')
-      editDialogModal.VerifyChatTurnIsAnExactMatch('Hello [$name]', 2, 1)
+      train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+      train.SelectChatTurnStartsWith('Hello')
+      train.VerifyErrorMessage('Action is unavailable')
+      train.VerifyChatTurnIsAnExactMatch('Hello [$name]', 2, 1)
     })
 
     it('Should fix the user turn that caused the error', () => {
-      editDialogModal.SelectChatTurnExactMatch('Joe')
-      editDialogModal.LabelTextAsEntity('Joe', 'name')
-      editDialogModal.ClickSubmitChangesButton()
+      train.SelectChatTurnExactMatch('Joe')
+      train.LabelTextAsEntity('Joe', 'name')
+      train.ClickSubmitChangesButton()
     })
 
     it('Should verify that there are no more errors', () => {
-      editDialogModal.VerifyNoErrorMessage()
-      editDialogModal.ClickSaveCloseButton()
+      train.VerifyNoErrorMessage()
+      train.ClickSaveCloseButton()
       modelPage.VerifyNoIncidentTriangleOnPage()
     })
   })
