+++ cypress/integration/ErrorHandling/TwoConsecutiveUserInput.spec.js
@@ -6,7 +6,6 @@
 import * as models from '../../support/Models'
 import * as modelPage from '../../support/components/ModelPage'
 import * as train from '../../support/Train'
-import * as editDialogModal from '../../support/components/EditDialogModal'
 import * as common from '../../support/Common'
 
 describe('ErrorHandling', () => {
@@ -18,12 +17,12 @@ describe('ErrorHandling', () => {
     modelPage.VerifyNoIncidentTriangleOnPage()
 
     train.EditTraining('Hey', 'world peace', "Sorry $name, I can't help you get $want")
-    editDialogModal.InsertUserInputAfter('Sam', 'InsertedText')
-    editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
-    editDialogModal.SelectChatTurnExactMatch('Sam')
-    editDialogModal.VerifyErrorMessage('Two consecutive User Inputs')
+    train.InsertUserInputAfter('Sam', 'InsertedText')
+    train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+    train.SelectChatTurnExactMatch('Sam')
+    train.VerifyErrorMessage('Two consecutive User Inputs')
 
-    editDialogModal.ClickSaveCloseButton()
+    train.ClickSaveCloseButton()
 
     modelPage.VerifyIncidentTriangleForTrainDialogs()
     train.VerifyIncidentTriangleFoundInTrainDialogsGrid(`Hey`, 'world peace', "Sorry $name, I can't help you get $want")
@@ -31,14 +30,14 @@ describe('ErrorHandling', () => {
     // - - - Open the same Train Dialog, validate and fix the errors. - - -
 
     train.EditTraining(`Hey`, 'world peace', "Sorry $name, I can't help you get $want")
-    editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
-    editDialogModal.SelectChatTurnExactMatch('Sam')
-    editDialogModal.VerifyErrorMessage('Two consecutive User Inputs')
-    editDialogModal.SelectChatTurnExactMatch('InsertedText')
-    editDialogModal.ClickDeleteChatTurn()
-    editDialogModal.VerifyNoErrorMessage()
+    train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+    train.SelectChatTurnExactMatch('Sam')
+    train.VerifyErrorMessage('Two consecutive User Inputs')
+    train.SelectChatTurnExactMatch('InsertedText')
+    train.ClickDeleteChatTurn()
+    train.VerifyNoErrorMessage()
 
-    editDialogModal.ClickSaveCloseButton()
+    train.ClickSaveCloseButton()
     modelPage.VerifyNoIncidentTriangleOnPage()
   })
 })
