+++ cypress/integration/ErrorHandling/WaitNonWait.spec.js
@@ -6,7 +6,6 @@
 import * as models from '../../support/Models'
 import * as modelPage from '../../support/components/ModelPage'
 import * as train from '../../support/Train'
-import * as editDialogModal from '../../support/components/EditDialogModal'
 import * as common from '../../support/Common'
 import * as helpers from '../../support/Helpers'
 
@@ -24,23 +23,23 @@ describe('Wait Non Wait Error Handling', () => {
   context('Create Errors', () => {
     it('Deletes a user turn to create an error, verifies expected (and no other) error messages come up', () => {
       train.EditTraining('Duck', 'Fish', common.fishJustSwim)
-      editDialogModal.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike)
-      editDialogModal.ClickDeleteChatTurn()
-      editDialogModal.VerifyErrorMessage(common.userInputFollowsNonWaitErrorMessage)
+      train.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike)
+      train.ClickDeleteChatTurn()
+      train.VerifyErrorMessage(common.userInputFollowsNonWaitErrorMessage)
 
       Validations(1)
     })
 
     it('Inserts an extra Bot response, verifies new error message and other error message remain', () => {
-      editDialogModal.InsertBotResponseAfter('Duck', common.whichAnimalWouldYouLike)
-      editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+      train.InsertBotResponseAfter('Duck', common.whichAnimalWouldYouLike)
+      train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
 
       Validations(2)
     })
 
     it('Inserts another Bot response, verifies new error message and other error messages remain', () => {
-      editDialogModal.InsertBotResponseAfter('Fish', common.whichAnimalWouldYouLike)
-      editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+      train.InsertBotResponseAfter('Fish', common.whichAnimalWouldYouLike)
+      train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
 
       Validations(3)
     })
@@ -48,14 +47,14 @@ describe('Wait Non Wait Error Handling', () => {
 
   context(`Save & Validate Training Errors`, () => {
     it('Saves the Training with Errors, verifies Model page and Train Dialog grid shows an error', () => {
-      editDialogModal.ClickSaveCloseButton()
+      train.ClickSaveCloseButton()
       modelPage.VerifyIncidentTriangleForTrainDialogs()
       train.VerifyIncidentTriangleFoundInTrainDialogsGrid(`Duck`, 'Fish', common.fishJustSwim)
     })
 
     it('Re-opens the Training and validates all error messages remain', () => {
       train.EditTraining(`Duck`, 'Fish', common.fishJustSwim)
-      editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+      train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
 
       Validations(3)
     })
@@ -63,29 +62,29 @@ describe('Wait Non Wait Error Handling', () => {
 
   context('Correct the Errors', () => {
     it('Deletes one of the error causing turns, verifies an error went away & other error remained', () => {
-      editDialogModal.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike, 1)
-      editDialogModal.ClickDeleteChatTurn()
-      editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+      train.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike, 1)
+      train.ClickDeleteChatTurn()
+      train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
 
       Validations(2)
     })
 
     it('Deletes another error causing turn, verifies another error went away & still one error remains', () => {
-      editDialogModal.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike)
-      editDialogModal.ClickDeleteChatTurn()
-      editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+      train.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike)
+      train.ClickDeleteChatTurn()
+      train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
 
       Validations(1)
     })
 
     it('Re-inserts the original user turn, verifies all errors are gone in dialog and on Model page', () => {
-      editDialogModal.SelectChatTurnExactMatch('Fish')
-      editDialogModal.VerifyErrorMessage(common.userInputFollowsNonWaitErrorMessage)
+      train.SelectChatTurnExactMatch('Fish')
+      train.VerifyErrorMessage(common.userInputFollowsNonWaitErrorMessage)
 
-      editDialogModal.InsertBotResponseAfter(common.ducksSayQuack, common.whichAnimalWouldYouLike)
-      editDialogModal.VerifyNoErrorMessage()
+      train.InsertBotResponseAfter(common.ducksSayQuack, common.whichAnimalWouldYouLike)
+      train.VerifyNoErrorMessage()
 
-      editDialogModal.ClickSaveCloseButton()
+      train.ClickSaveCloseButton()
       modelPage.VerifyNoIncidentTriangleOnPage()
     })
   })
@@ -93,29 +92,29 @@ describe('Wait Non Wait Error Handling', () => {
 
 function Validations(errCount) {
   cy.ConLog(`Validations(${errCount})`, `Start`)
-  editDialogModal.SelectChatTurnExactMatch('Duck')
-  editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+  train.SelectChatTurnExactMatch('Duck')
+  train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
 
   if (errCount > 1) {
-    editDialogModal.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike)
-    editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+    train.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike)
+    train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
   }
 
-  editDialogModal.SelectChatTurnExactMatch(common.ducksSayQuack)
-  if (errCount == 1) editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
-  else editDialogModal.VerifyErrorMessage(common.actionFollowsWaitActionErrorMessage)
+  train.SelectChatTurnExactMatch(common.ducksSayQuack)
+  if (errCount == 1) train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+  else train.VerifyErrorMessage(common.actionFollowsWaitActionErrorMessage)
 
-  editDialogModal.SelectChatTurnExactMatch('Fish')
-  editDialogModal.VerifyErrorMessage(common.userInputFollowsNonWaitErrorMessage)
+  train.SelectChatTurnExactMatch('Fish')
+  train.VerifyErrorMessage(common.userInputFollowsNonWaitErrorMessage)
 
   if (errCount > 2) {
-    editDialogModal.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike, 1)
-    editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+    train.SelectChatTurnExactMatch(common.whichAnimalWouldYouLike, 1)
+    train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
   }
 
-  editDialogModal.SelectChatTurnExactMatch(common.fishJustSwim)
-  if (errCount < 3) editDialogModal.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
-  else editDialogModal.VerifyErrorMessage(common.actionFollowsWaitActionErrorMessage)
+  train.SelectChatTurnExactMatch(common.fishJustSwim)
+  if (errCount < 3) train.VerifyErrorMessage(common.trainDialogHasErrorsMessage)
+  else train.VerifyErrorMessage(common.actionFollowsWaitActionErrorMessage)
 
   cy.ConLog(`Validations(${errCount})`, `End`)
 }
