+++ cypress/integration/Train/WaitVsNoWaitActions.spec.js
@@ -7,7 +7,6 @@ import * as models from '../../support/Models'
 import * as modelPage from '../../support/components/ModelPage'
 import * as scorerModal from '../../support/components/ScorerModal'
 import * as train from '../../support/Train'
-import * as editDialogModal from '../../support/components/EditDialogModal'
 import * as common from '../../support/Common'
 
 describe('Train', () => {
@@ -20,14 +19,14 @@ describe('Train', () => {
     train.CreateNewTrainDialog()
 
     train.TypeYourMessage('Hello')
-    editDialogModal.ClickScoreActionsButton()
+    train.ClickScoreActionsButton()
     scorerModal.VerifyContainsEnabledAction('Which animal would you like?')
     scorerModal.VerifyContainsEnabledAction('Cows say moo!')
     scorerModal.VerifyContainsEnabledAction(common.ducksSayQuack)
     train.SelectAction('Which animal would you like?')
 
     train.TypeYourMessage('Cow')
-    editDialogModal.ClickScoreActionsButton()
+    train.ClickScoreActionsButton()
     scorerModal.VerifyContainsEnabledAction('Which animal would you like?')
     scorerModal.VerifyContainsEnabledAction('Cows say moo!')
     scorerModal.VerifyContainsEnabledAction(common.ducksSayQuack)
@@ -36,7 +35,7 @@ describe('Train', () => {
     train.SelectAction('Which animal would you like?')
 
     train.TypeYourMessage('Duck')
-    editDialogModal.ClickScoreActionsButton()
+    train.ClickScoreActionsButton()
     scorerModal.VerifyContainsEnabledAction('Which animal would you like?')
     scorerModal.VerifyContainsEnabledAction('Cows say moo!')
     scorerModal.VerifyContainsEnabledAction(common.ducksSayQuack)
